Move nrunner standalone runners to a new runners dir
With the possibility of having new runners added to the nrunner
architecture, let's move the standalone runners to a specific directory,
just like the spawners' structure.
