+++ avocado/core/nrunner.py
@@ -255,10 +255,10 @@ class Runnable:
         # Looking for the file only avoids an attempt to load the module
         # and should be a lot faster
         core_dir = os.path.dirname(os.path.abspath(__file__))
-        module_name = "nrunner_%s" % self.kind.replace('-', '_')
+        module_name = self.kind.replace('-', '_')
         module_filename = '%s.py' % module_name
-        if os.path.exists(os.path.join(core_dir, module_filename)):
-            full_module_name = 'avocado.core.%s' % module_name
+        if os.path.exists(os.path.join(core_dir, 'runners', module_filename)):
+            full_module_name = 'avocado.core.runners.%s' % module_name
             candidate_cmd = [sys.executable, '-m', full_module_name]
             if self.is_kind_supported_by_runner_command(candidate_cmd):
                 runners_registry[self.kind] = candidate_cmd
